requirements.txt
flask
flask-bootstrap
flask-wtf
flask-testing
blinker
firebase-admin
flask-login
ipykernel==6.7.0
ipython==8.0.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.22.0
pandas==1.3.5
scikit-learn==1.0.2
scipy==1.7.3
sklearn==0.0
stack-data==0.1.3
pyrtlsdr==0.2.92
pathlib